further trim ui to its bare ui state, and actually render in draw.rs
